+++ docs/make.jl
@@ -1,4 +1,4 @@
-using Documenter
+import Documenter
 using QuantumCircuitOpt
 using DocumenterTools: Themes
 
@@ -17,17 +17,18 @@ Themes.compile(
     joinpath(@__DIR__, "src/assets/themes/dark.scss"),
 )
 
-makedocs(
+Documenter.makedocs(
     modules = [QuantumCircuitOpt],
     sitename = "QuantumCircuitOpt.jl",
     format = Documenter.HTML(mathengine = Documenter.MathJax(), 
-                             assets = [asset("https://fonts.googleapis.com/css?family=Montserrat|Source+Code+Pro&display=swap", class=:css),
+                             assets = [Documenter.asset("https://fonts.googleapis.com/css?family=Montserrat|Source+Code+Pro&display=swap", class=:css),
                                              "assets/extra_styles.css"],
                              prettyurls = get(ENV, "CI", nothing) == "true",
                              sidebar_sitename=false
                              ),
     # strict = true,
     authors = "Harsha Nagarajan",
+    # checkdocs = :none,
     pages = [
         "Introduction" => "index.md",
         "Quick Start guide" => "quickguide.md",
@@ -41,7 +42,7 @@ makedocs(
     ],
 )
 
-deploydocs(
+Documenter.deploydocs(
     repo = "github.com/harshangrjn/QuantumCircuitOpt.jl.git",
     push_preview = true
 )