+++ examples/run_examples.jl
@@ -3,7 +3,7 @@ using LinearAlgebra
 
 using JuMP
 using Gurobi
-using CPLEX
+# using CPLEX
 # using HiGHS
 
 include("optimizers.jl")
@@ -11,12 +11,12 @@ include("optimizers.jl")
 include("parametrized_gates.jl")
 include("decompose_all_gates.jl")
 
-# decompose_gates = ["QFT3"]
+decompose_gates = ["minimize_T_gate"]
 
 #----------------------------------------------#
 #      Quantum Circuit Optimization model      #
 #----------------------------------------------#
-qcopt_optimizer = get_gurobi(solver_log = true)
+qcopt_optimizer = get_gurobi(solver_log = false)
 
 result = Dict{String,Any}()
 times = zeros(length(decompose_gates), 1)